Issue #8 Move window to bottom of screen.
The window is anchored to bottom right, so can only be in bottom or right
of screen.
Update set_win_tiles to use the new Y location.
The transparency in window is ONLY for sprites, not background.
This means wherever the window overlaps, it will cover the background
-e 
Original/build commit ID: 32ca9e3b447cf822ad508363087a2dc40596af6f