Issue #91 Remove potential bug in addition of unitialised tile_data_offset variable.
tile_data_offset was originally set based on an old variable of the
menu_config, which would specify a base 'offset' tile and all tile values
were indexes from this value.
This had since been removed, but we still seem to have a global defined
for holding the offset and was still being added to tile_index, which
could have resulted in any amount being added to the tile_index, depending
on what the value happened to be.
-e 
Original/build commit ID: edf0836b64368946bba7100d01ee6251632ef091