Issue #25 Update set_tile_direction to handle 2x2 sprites.
This involves setting the second row (or column) of sprite tiles to a
different set of 3 tiles, based on direction.
Update move_ai_sprite to no longer attempt to update direction of sprites
that have a move speed of 0, which is now treated as disabling the
movement. The sprite direction should be set for all tiles, to ensure that
they are set to the correct static direction by set_sprite_direction.
Fix base tile and direction of house car sprite
-e 
Original/build commit ID: 7270abe8256a94fd9118024fbd75a891b0fb9d87